Integrate: [ 29436] Upgrade to Devel::PPPort 3.10_01
[ 29437]
Upgrade to Devel::PPPort 3.10_02

[ 30280]
Upgrade to Devel::PPPort 3.11
p4raw-link: @30280 on //depot/perl: d2dacc4f735315b0cceecc21aa724d1bd4b2b797
p4raw-link: @29437 on //depot/perl: 91de150c553e264178070aa0dcef035c37a19eb2
p4raw-link: @29436 on //depot/perl: cac25305bdd103d726ae355802b63e8a67132aef

p4raw-id: //depot/maint-5.8/perl@30287
p4raw-branched: from //depot/perl@30286 'branch in'
	ext/Devel/PPPort/parts/base/5009005
	ext/Devel/PPPort/parts/todo/5009005
p4raw-integrated: from //depot/perl@30286 'copy in'
	ext/Devel/PPPort/README ext/Devel/PPPort/module2.c
	ext/Devel/PPPort/module3.c ext/Devel/PPPort/parts/inc/MY_CXT
	ext/Devel/PPPort/parts/inc/SvPV ext/Devel/PPPort/parts/inc/call
	ext/Devel/PPPort/parts/inc/cop
	ext/Devel/PPPort/parts/inc/exception
	ext/Devel/PPPort/parts/inc/format
	ext/Devel/PPPort/parts/inc/grok
	ext/Devel/PPPort/parts/inc/limits
	ext/Devel/PPPort/parts/inc/mPUSH
	ext/Devel/PPPort/parts/inc/newCONSTSUB
	ext/Devel/PPPort/parts/inc/newRV
	ext/Devel/PPPort/parts/inc/sv_xpvf
	ext/Devel/PPPort/parts/inc/uv
	ext/Devel/PPPort/parts/inc/version ext/Devel/PPPort/typemap
	(@26866..) ext/Devel/PPPort/parts/inc/Sv_set
	ext/Devel/PPPort/parts/inc/magic
	ext/Devel/PPPort/parts/inc/snprintf
	ext/Devel/PPPort/parts/inc/warn (@28259..)
	ext/Devel/PPPort/parts/inc/SvREFCNT
	ext/Devel/PPPort/parts/inc/memory
	ext/Devel/PPPort/parts/inc/pvs (@28267..)
	ext/Devel/PPPort/apicheck_c.PL
	ext/Devel/PPPort/devel/buildperl.pl
	ext/Devel/PPPort/parts/base/5005000
	ext/Devel/PPPort/parts/base/5007002
	ext/Devel/PPPort/parts/base/5009001
	ext/Devel/PPPort/parts/todo/5009001 (@28332..)
	ext/Devel/PPPort/t/MY_CXT.t ext/Devel/PPPort/t/SvPV.t
	ext/Devel/PPPort/t/SvREFCNT.t ext/Devel/PPPort/t/Sv_set.t
	ext/Devel/PPPort/t/call.t ext/Devel/PPPort/t/cop.t
	ext/Devel/PPPort/t/exception.t ext/Devel/PPPort/t/grok.t
	ext/Devel/PPPort/t/limits.t ext/Devel/PPPort/t/mPUSH.t
	ext/Devel/PPPort/t/magic.t ext/Devel/PPPort/t/memory.t
	ext/Devel/PPPort/t/misc.t ext/Devel/PPPort/t/newCONSTSUB.t
	ext/Devel/PPPort/t/newRV.t ext/Devel/PPPort/t/pvs.t
	ext/Devel/PPPort/t/snprintf.t ext/Devel/PPPort/t/sv_xpvf.t
	ext/Devel/PPPort/t/threads.t ext/Devel/PPPort/t/uv.t
	ext/Devel/PPPort/t/variables.t ext/Devel/PPPort/t/warn.t
	(@28424..) ext/Devel/PPPort/parts/inc/ppphbin
	ext/Devel/PPPort/parts/inc/ppphdoc
	ext/Devel/PPPort/parts/inc/ppphtest
	ext/Devel/PPPort/t/ppphtest.t (@28474..)
	ext/Devel/PPPort/Makefile.PL ext/Devel/PPPort/parts/inc/podtest
	ext/Devel/PPPort/t/podtest.t (@28507..)
	ext/Devel/PPPort/HACKERS ext/Devel/PPPort/parts/ppptools.pl
	(@28605..) ext/Devel/PPPort/parts/apicheck.pl
	ext/Devel/PPPort/parts/base/5009004
	ext/Devel/PPPort/parts/embed.fnc
	ext/Devel/PPPort/parts/inc/strlfuncs
	ext/Devel/PPPort/parts/todo/5009004
	ext/Devel/PPPort/t/strlfuncs.t (@28616..)
	ext/Devel/PPPort/PPPort_xs.PL ext/Devel/PPPort/TODO
	ext/Devel/PPPort/parts/inc/misc
	ext/Devel/PPPort/parts/inc/threads
	ext/Devel/PPPort/parts/inc/variables (@29436..)
	ext/Devel/PPPort/PPPort_pm.PL ext/Devel/PPPort/soak (@29437..)
p4raw-integrated: from //depot/perl@29437 'ignore'
	ext/Devel/PPPort/Changes (@29436..) 'merge in' MANIFEST
	(@29432..)
p4raw-integrated: from //depot/perl@29436 'ignore'
	ext/Devel/PPPort/mktests.PL (@28605..)

+++ ext/Devel/PPPort/devel/buildperl.pl
@@ -5,9 +5,9 @@
 #
 ################################################################################
 #
-#  $Revision: 9 $
+#  $Revision: 10 $
 #  $Author: mhx $
-#  $Date: 2006/05/28 20:42:53 +0200 $
+#  $Date: 2006/12/02 09:58:34 +0100 $
 #
 ################################################################################
 #
@@ -39,6 +39,7 @@ my %opt = (
   force   => 0,
   test    => 0,
   install => 1,
+  'test-archives' => 0,
 );
 
 my %config = (
@@ -110,6 +111,7 @@ GetOptions(\%opt, qw(
   force
   test
   install!
+  test-archives+
 )) or pod2usage(2);
 
 if (exists $opt{config}) {
@@ -137,6 +139,24 @@ if (exists $opt{perl}) {
 }
 else {
   @perls = sort keys %perl;
+}
+
+if ($opt{'test-archives'}) {
+  my $test = 'test';
+  my $cwd = cwd;
+  -d $test or mkpath($test);
+  chdir $test or die "chdir $test: $!\n";
+  for my $perl (@perls) {
+    eval {
+      my $d = extract_source($perl{$perl});
+      rmtree($d) if -e $d;
+    };
+    warn $@ if $@;
+  }
+  chdir $cwd or die "chdir $cwd: $!\n";
+  print STDERR "cleaning up\n";
+  rmtree($test);
+  exit 0;
 }
 
 my %current;
@@ -211,7 +231,8 @@ sub extract_source
 {
   my $perl = shift;
 
-  print "reading $perl->{source}\n";
+  my $what = $opt{'test-archives'} ? 'test' : 'read';
+  print "${what}ing $perl->{source}\n";
 
   my $target;
 
@@ -222,17 +243,19 @@ sub extract_source
     $target = $t;
   }
 
-  if (-d $target) {
-    print "removing old build directory $target\n";
-    rmtree($target);
+  if ($opt{'test-archives'} == 0 || $opt{'test-archives'} > 1) {
+    if (-d $target) {
+      print "removing old build directory $target\n";
+      rmtree($target);
+    }
+
+    print "extracting $perl->{source}\n";
+
+    Archive::Tar->extract_archive($perl->{source})
+        or die "extract failed: " . Archive::Tar->error() . "\n";
+
+    -d $target or die "oooops, $target not found\n";
   }
-
-  print "extracting $perl->{source}\n";
-
-  Archive::Tar->extract_archive($perl->{source})
-      or die "extract failed: " . Archive::Tar->error() . "\n";
-
-  -d $target or die "oooops, $target not found\n";
 
   return $target;
 }
